+++ components/edgecontroller/timeseries_processor.go
@@ -22,8 +22,9 @@ import (
 	"encoding/hex"
 	"time"
 
-	agentv1 "github.com/leansignal/leansignal-agent/proto/gen/leansignal/agent/v1"
 	"go.uber.org/zap"
+
+	agentv1 "github.com/leansignal/leansignal-agent/proto/gen/leansignal/agent/v1"
 )
 
 // Constants for the timeseries processor